+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - locate_columns method and col_nums global variable to consolidate information
 - GUI_DEBUG global variable in layout.py to pass update_color during debugging
 - key in the progress bar space
+- key appears only when output page is selected
 
 ### Fixed
 - error message about manual synonyms column input corrected
 - status bar will always align with bottom of window (added anchor=s)
+- ignores capitalization of column titles in Excel file
+- openpyxl library updated to 2.4.5 for better compatibility with Excel
 
 ### Removed
 - wordcloud debugging print statements that locate DroidSansMono.ttf